Dialouge Trigger, Dialogue System Del
