+++ Assets/JA/Enums.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 089e475445450554e888240ec938f3ef
+guid: ed2effd2dd47cb242a96524c40dc47e5
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
