+++ Assets/JA/Scripts/Class.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 02a4f9d10690b8e4c872c1637e2c45a0
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
