+++ Assets/JA/Scripts/Class/Customer.cs
@@ -1,0 +1,10 @@
+using System;
+
+[Serializable]
+public class Customer
+{
+    public string name;
+    public int id;
+    public Perfume currentPerfume;
+    
+}
