+++ Assets/JA/Scripts/Class/Customer.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 677e30db9bfd4d3418d4e2e71db07902
+guid: ed76d50c1d6515b43984ad766f1a2d82
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
