+++ Assets/JA/Scripts/Class/Perfume.cs
@@ -1,0 +1,10 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+public class Perfume
+{
+    public string bassNotes;
+    public string MiddleNotes;
+    public string TopNotes;
+}
