+++ Assets/JA/Scripts/Class/Perfume.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: c5e1ea66fd691204392434d64c825044
+guid: 020a2d249fa5da44a88ab9df3ace332f
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
