--- Assets/script/Dialogue/imsi/Dialogue.cs
@@ -1,10 +1,0 @@
-using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-
-[System.Serializable]
-public class Dialogue
-{
-    public List<DialogueBuffer> BuyerSentences;
-    public List<DialogueBuffer> SellerSentences;
-}
