--- Assets/script/Dialogue/imsi/DialogueTrigger.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 26e3ddce78805a3468ffe12cf9060d5c
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
